fix: Resolved issue when viewing on Safari
This commit resolved the error that occurs when the application is
running on the Safari browser. Changes were made to the application's
MVVM such that the Knockout binding does not occur until the Google Map
is completely initialized. Files were also relocated to improve their
organization.

The error which this commit resolves pertains to the control flow of
the application when running on Safari. The application's MVVM was
being run before the complete initialization of the Google Map. Thus,
the MVVM could not render the views due to missing dependencies.