feat: Added image viewer
This commit adds a new feature to view images in a lightbox gallery
using the Magnific Popup jQuery plugin. Images clicked in each info-
window will now expand into the lightbox with all the other images
in the gallery. A minor stylistic change was also made to fit the
dropdown for difficulty selection inside the header box.

This feature addresses a problem with viewing images of a hiking
trail. Each image would load up in a new window, forcing the user
to have to go back in order to return to the application.